Shorten the dashed lines to a more standard 80 column width
The comments with dashed lines & headings nicely separate files in
sections, but they were 90 characters line.
This is a bit annoying when an editor is configured to have width of
80 characters. This is a common choice, as it often is just enough to
comfortably have two columns side-by-side.
(Actually I stopped the lines to 78 characters, because sometimes
column width is set a little shorted still)